+++ compiler-compat/k250_dev_498/src/main/kotlin/dev/zacsweers/metro/compiler/compat/k250_dev_498/CompatContextImpl.kt
@@ -4,35 +4,12 @@ package dev.zacsweers.metro.compiler.compat.k250_dev_498
 
 import dev.zacsweers.metro.compiler.compat.CompatContext
 import dev.zacsweers.metro.compiler.compat.k2420_beta2.CompatContextImpl as DelegateType
-import org.jetbrains.kotlin.fir.expressions.FirResolvedQualifier
-import org.jetbrains.kotlin.fir.expressions.builder.buildResolvedQualifier
-import org.jetbrains.kotlin.fir.symbols.impl.FirClassLikeSymbol
-import org.jetbrains.kotlin.fir.types.ConeKotlinType
 import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
 import org.jetbrains.kotlin.ir.declarations.IrPackageFragment
 import org.jetbrains.kotlin.ir.declarations.createEmptyExternalPackageFragment as createEmptyExternalPackageFragmentNative
-import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.FqName
 
 public class CompatContextImpl : CompatContext by DelegateType() {
-  override fun buildResolvedQualifierCompat(
-    classId: ClassId,
-    classSymbol: FirClassLikeSymbol<*>,
-    classType: ConeKotlinType,
-  ): FirResolvedQualifier {
-    return buildResolvedQualifier {
-      packageFqName = classId.packageFqName
-      relativeClassFqName = classId.relativeClassName
-      qualifierSymbol = classSymbol
-      resolvedToCompanionObject = false
-      explicitParent = buildResolvedQualifier {
-        packageFqName = classId.packageFqName
-        resolvedToCompanionObject = false
-      }
-      coneTypeOrNull = classType
-    }
-  }
-
   override fun IrModuleFragment.createEmptyExternalPackageFragmentCompat(
     packageName: String
   ): IrPackageFragment {
